+++ src/dtos/term.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Deserialize)]
 pub struct CreateTermDto {
-    pub project_id: i32,
+    pub project_id: u64,
     pub source_term: String,
     pub target_term: String,
     pub language_code: String,
@@ -11,8 +11,8 @@ pub struct CreateTermDto {
 
 #[derive(Debug, Serialize)]
 pub struct TermDto {
-    pub id: i32,
-    pub project_id: i32,
+    pub id: u64,
+    pub project_id: u64,
     pub source_term: String,
     pub target_term: String,
     pub language_code: String,
